Sort out attachment api
Never do any magic encoding or decoding of attachment content. You are
passing a str, bytes or stream of bytes into the attachment content. This
means you must encode to utf8 before hand if saving a unicode string.

Similarly never do any magic decoding. The return value of get_attachment()
is going to be a str or a stream of bytes. Never unicode!!